requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
alembic==1.16.4
celery==5.5.3
redis==6.2.0
python-telegram-bot==22.2
python-multipart==0.0.20
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
pydantic==2.11.7
pydantic-settings==2.10.1
pytest==8.4.1
pytest-asyncio==1.1.0
httpx==0.28.1
faker==37.4.2
